requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
click==8.1.3
cvxopt==1.3.0
cycler==0.11.0
Cython==0.29.32
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-bootstrap-templates==1.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
executing==0.9.0
Flask==2.2.2
fonttools==4.34.4
graphviz==0.20.1
h5py==3.7.0
importlib-metadata==6.0.0
infinity==1.5
intervaltree==3.1.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
kiwisolver==1.4.4
llvmlite==0.39.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mpmath==1.2.1
networkx==2.8.5
numba==0.56.0
numpy==1.22.4
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.11.0
pm4py==2.2.24
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pydotplus==2.0.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyvis==0.2.1
PyYAML==6.0
scikit-learn==1.1.1
scipy==1.8.1
shapely==2.0.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.3.0
StringDist==1.0.9
sympy==1.10.1
tenacity==8.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
traces==0.6.0
traitlets==5.3.0
tslearn==0.5.2
wcwidth==0.2.5
Werkzeug==2.2.2
xes==1.3
zipp==3.11.0
